organize files a bit better based on app and common folders
