+++ src/common/constants/lofi-tracks.ts
@@ -1,28 +1,28 @@
-import track1 from '../assets/music/lofi/01 No Spirit - Glowing lights (Kupla Master) (1).mp3';
-import track2 from '../assets/music/lofi/02 dryhope - somewhere else (Kupla Master) (1).mp3';
-import track3 from '../assets/music/lofi/03 Eisu - sheets (Kupla Master) (1).mp3';
-import track4 from '../assets/music/lofi/04 Casiio _ Sleepermane - Phantasm (Kupla Master) (1).mp3';
-import track5 from '../assets/music/lofi/05 Project AER x WYS - after sunset (Kupla Master) (1).mp3';
-import track6 from '../assets/music/lofi/06 Tenno - Luna (Kupla Master) (1).mp3';
-import track7 from '../assets/music/lofi/07 Sebastian Kamae _ Intoku - Lucid (Kupla Master) (1).mp3';
-import track8 from '../assets/music/lofi/08 epektase x j_san - contemplation (Kupla Master) (1).mp3';
-import track9 from '../assets/music/lofi/09 Casiio _ Sleepermane - Signals (Kupla Master) (1).mp3';
-import track10 from '../assets/music/lofi/10 Phlocalyst _ Living Room - Feelings (Kupla Master) (1).mp3';
-import track11 from '../assets/music/lofi/11 Pandrezz - Blind Forest (Kupla Master) (1).mp3';
-import track12 from '../assets/music/lofi/12 SwuM x chief. - Show Me How (Kupla Master) (1).mp3';
-import track13 from '../assets/music/lofi/13 Comodo - Numb (Kupla Master) (1).mp3';
-import track14 from '../assets/music/lofi/14 TABAL x mell-ø - Weightless (Kupla Master) (1).mp3';
-import track15 from '../assets/music/lofi/15 Ambulo - Night Bus (Kupla Master) (1).mp3';
-import track16 from '../assets/music/lofi/16 Kanisan x Frad - Fireflies (Kupla Master) (1).mp3';
-import track17 from '../assets/music/lofi/17 H.1 - Timeless (Kupla Master) (1).mp3';
-import track18 from '../assets/music/lofi/18 Yasumu - Recharge (Kupla Master) (1).mp3';
-import track19 from '../assets/music/lofi/19 cxlt. - fade away (Kupla Master) (1).mp3';
-import track20 from '../assets/music/lofi/20 Mondo Loops x Kanisan - Escape Route (Kupla Master) (1).mp3';
-import track21 from '../assets/music/lofi/21 towerz - fateful slumber (Kupla Master) (1).mp3';
-import track22 from '../assets/music/lofi/22 brillion. - Crescent (Kupla Master) (1).mp3';
-import track23 from '../assets/music/lofi/23 less.people - persist (Kupla Master2).mp3';
-import track24 from '../assets/music/lofi/24 Mondoloops - Drifting Far Away (Kupla Master) (1).mp3';
-import threeAmArtwork from '../assets/3am-session-artwork.png';
+import track1 from '@/assets/music/lofi/01 No Spirit - Glowing lights (Kupla Master) (1).mp3';
+import track2 from '@/assets/music/lofi/02 dryhope - somewhere else (Kupla Master) (1).mp3';
+import track3 from '@/assets/music/lofi/03 Eisu - sheets (Kupla Master) (1).mp3';
+import track4 from '@/assets/music/lofi/04 Casiio _ Sleepermane - Phantasm (Kupla Master) (1).mp3';
+import track5 from '@/assets/music/lofi/05 Project AER x WYS - after sunset (Kupla Master) (1).mp3';
+import track6 from '@/assets/music/lofi/06 Tenno - Luna (Kupla Master) (1).mp3';
+import track7 from '@/assets/music/lofi/07 Sebastian Kamae _ Intoku - Lucid (Kupla Master) (1).mp3';
+import track8 from '@/assets/music/lofi/08 epektase x j_san - contemplation (Kupla Master) (1).mp3';
+import track9 from '@/assets/music/lofi/09 Casiio _ Sleepermane - Signals (Kupla Master) (1).mp3';
+import track10 from '@/assets/music/lofi/10 Phlocalyst _ Living Room - Feelings (Kupla Master) (1).mp3';
+import track11 from '@/assets/music/lofi/11 Pandrezz - Blind Forest (Kupla Master) (1).mp3';
+import track12 from '@/assets/music/lofi/12 SwuM x chief. - Show Me How (Kupla Master) (1).mp3';
+import track13 from '@/assets/music/lofi/13 Comodo - Numb (Kupla Master) (1).mp3';
+import track14 from '@/assets/music/lofi/14 TABAL x mell-ø - Weightless (Kupla Master) (1).mp3';
+import track15 from '@/assets/music/lofi/15 Ambulo - Night Bus (Kupla Master) (1).mp3';
+import track16 from '@/assets/music/lofi/16 Kanisan x Frad - Fireflies (Kupla Master) (1).mp3';
+import track17 from '@/assets/music/lofi/17 H.1 - Timeless (Kupla Master) (1).mp3';
+import track18 from '@/assets/music/lofi/18 Yasumu - Recharge (Kupla Master) (1).mp3';
+import track19 from '@/assets/music/lofi/19 cxlt. - fade away (Kupla Master) (1).mp3';
+import track20 from '@/assets/music/lofi/20 Mondo Loops x Kanisan - Escape Route (Kupla Master) (1).mp3';
+import track21 from '@/assets/music/lofi/21 towerz - fateful slumber (Kupla Master) (1).mp3';
+import track22 from '@/assets/music/lofi/22 brillion. - Crescent (Kupla Master) (1).mp3';
+import track23 from '@/assets/music/lofi/23 less.people - persist (Kupla Master2).mp3';
+import track24 from '@/assets/music/lofi/24 Mondoloops - Drifting Far Away (Kupla Master) (1).mp3';
+import threeAmArtwork from '@/assets/3am-session-artwork.png';
 
 export const LOFI_PLAYER_TRACKS = [
   {
