+++ src/common/utils/timerHelpers.ts
