--- src/features/index.ts
@@ -1,3 +1,0 @@
-export * from './timer';
-export * from './todolist';
-export * from './buttons';
