+++ src/features/timer/PomoTimer.less
@@ -1,4 +1,4 @@
-@import '../../constants/styles.less';
+@import '../../common/constants/styles.less';
 
 .pomo-timer {
   display: flex;
