+++ src/features/todolist/Todo.less
@@ -1,4 +1,4 @@
-@import '../../constants/styles.less';
+@import '../../common/constants/styles.less';
 
 .todo {
   margin: 8px 0;
