+++ src/features/todolist/TodoForm.less
@@ -1,4 +1,4 @@
-@import '../../constants/styles.less';
+@import '../../common/constants/styles.less';
 
 .todo-form {
   display: flex;
