+++ src/features/todolist/TodoForm.tsx
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { Checkbox, TextField } from '@mui/material';
-import { TodoFormProps } from '@/types';
-import { TODO_FORM_HELPER_TEXT_FIELD, TODO_FORM_NAME_CHECKBOX, TODO_FORM_NAME_TEXTFIELD } from '@/constants';
+import { TodoFormProps } from '@/common/types';
+import { TODO_FORM_HELPER_TEXT_FIELD, TODO_FORM_NAME_CHECKBOX, TODO_FORM_NAME_TEXTFIELD } from '@/common/constants';
 import './TodoForm.less';
 
 export function TodoForm({ onCreateTodo }: TodoFormProps) {
