+++ src/features/todolist/redux/deleteCompleteTodos.ts
@@ -1,4 +1,4 @@
-import { TodosState } from '@/types';
+import { TodosState } from '@/common/types';
 
 export function deleteCompleteTodos(state: TodosState) {
   state.todos = state.todos.filter(({ isComplete }) => !isComplete);
