+++ src/features/todolist/redux/deleteTodo.ts
@@ -1,5 +1,5 @@
 import { PayloadAction } from '@reduxjs/toolkit';
-import { TodosState } from '../../../types';
+import { TodosState } from '@/common/types';
 
 export function deleteTodo(state: TodosState, action: PayloadAction<string>) {
   state.todos = state.todos.filter(({ id }) => id !== action.payload);
