+++ src/features/todolist/redux/errorLoadingTodos.ts
@@ -1,5 +1,5 @@
-import { FAILED } from '@/constants';
-import { TodosState } from '@/types';
+import { FAILED } from '@/common/constants';
+import { TodosState } from '@/common/types';
 
 export function errorLoadingTodos(state: TodosState) {
   state.loading = FAILED;
