+++ src/features/todolist/redux/getTodos.ts
@@ -1,4 +1,4 @@
-import { TodosState } from '../../../types';
+import { TodosState } from '@/common/types';
 
 export function getTodos(state: TodosState) {
   return state.todos;
