+++ src/features/todolist/redux/loadedTodos.ts
@@ -1,6 +1,6 @@
-import { SUCCEEDED } from '@/constants';
-import { TodoItem, TodosState } from '@/types';
 import { PayloadAction } from '@reduxjs/toolkit';
+import { SUCCEEDED } from '@/common/constants';
+import { TodoItem, TodosState } from '@/common/types';
 
 export function loadedTodos(state: TodosState, action: PayloadAction<TodoItem[]>) {
   state.todos = action.payload;
