+++ src/features/todolist/redux/loadingTodos.ts
@@ -1,5 +1,5 @@
-import { PENDING } from '@/constants';
-import { TodosState } from '@/types';
+import { PENDING } from '@/common/constants';
+import { TodosState } from '@/common/types';
 
 export function loadingTodos(state: TodosState) {
   state.loading = PENDING;
