+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { store } from '@/app/store';
-import App from './App.tsx';
+import App from '@/app/App.tsx';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
